add stdproc.multilook sub-module
+ add `stdproc.multilook` as WIP for version 2: start moving the low-level standardized processing functions into a new sub-module `stdproc`.

+ update `docs/api/module_hierarchy` for the new sub-module
